server/restate_services/python-hello-world/requirements.txt
hypercorn
restate_sdk==0.7.1
pydantic
SQLAlchemy~=2.0.0
asyncpg==0.29.0
python-dotenv~=1.0.0
openai>=1.78.0